style: clippy 1.98 findings in the new f16 tests and #858 test docs
unnecessary_cast in the new subnormal path (both float16.rs copies) and a
doc-list-indentation lint in regression_known_bugs.rs, where the #858 test
had also been inserted between the issue-15 test's doc comment and its
#[test] attribute; reordered.

+++ zenjxl-decoder/src/util/float16.rs
@@ -349,7 +349,7 @@ mod tests {
         if shift >= 32 {
             return sign; // far below the smallest subnormal
         }
-        let m = (full >> shift) as u32;
+        let m = full >> shift;
         let round_bit = (full >> (shift - 1)) & 1;
         let sticky = full & ((1u32 << (shift - 1)) - 1);
         let m = if round_bit == 1 && (sticky != 0 || (m & 1) == 1) {
